share_idx in ParSignedData changed to u8

+++ crates/dkg/src/signing.rs
@@ -107,7 +107,7 @@ pub enum SigningError {
 }
 
 /// Returns partially signed signatures of the lock hash.
-pub fn sign_lock_hash(share_idx: u64, shares: &[Share], hash: &[u8]) -> Result<ParSignedDataSet> {
+pub fn sign_lock_hash(share_idx: u8, shares: &[Share], hash: &[u8]) -> Result<ParSignedDataSet> {
     let mut set = ParSignedDataSet::new();
 
     for share in shares {
@@ -124,7 +124,7 @@ pub fn sign_lock_hash(share_idx: u64, shares: &[Share], hash: &[u8]) -> Result<P
 /// pubkey.
 pub fn sign_deposit_msgs(
     shares: &[Share],
-    share_idx: u64,
+    share_idx: u8,
     withdrawal_addresses: &[String],
     network_name: &str,
     amount: phase0::Gwei,
@@ -156,7 +156,7 @@ pub fn sign_deposit_msgs(
 /// Returns partially signed validator registrations keyed by validator pubkey.
 pub fn sign_validator_registrations(
     shares: &[Share],
-    share_idx: u64,
+    share_idx: u8,
     fee_recipients: &[String],
     gas_limit: u64,
     fork_version: &[u8],
@@ -216,7 +216,7 @@ pub(crate) async fn sign_and_agg_deposit_data(
     deposit_amounts: &[phase0::Gwei],
     compounding: bool,
 ) -> Result<Vec<Vec<phase0::DepositData>>> {
-    let share_idx = u64::try_from(node_idx.share_idx)?;
+    let share_idx = u8::try_from(node_idx.share_idx)?;
     let mut result = Vec::with_capacity(deposit_amounts.len());
 
     for (i, &amount) in deposit_amounts.iter().enumerate() {
@@ -260,7 +260,7 @@ pub(crate) async fn sign_and_agg_validator_registrations(
         gas_limit
     };
 
-    let share_idx = u64::try_from(node_idx.share_idx)?;
+    let share_idx = u8::try_from(node_idx.share_idx)?;
     let (set, msgs) = sign_validator_registrations(
         shares,
         share_idx,
@@ -347,7 +347,7 @@ pub(crate) async fn sign_and_aggregate_lock_hash(
         .cloned()
         .collect();
 
-    let share_idx = u64::try_from(node_idx.share_idx)?;
+    let share_idx = u8::try_from(node_idx.share_idx)?;
     let lock_hash_sig_set = sign_lock_hash(share_idx, &all_shares, &lock.lock_hash)?;
     let peer_sigs = exchanger.exchange(SIG_LOCK, lock_hash_sig_set).await?;
